submissions/0xa620a02c4cc2d20d/week4/backend/requirements.txt
sqlmodel>=0.0.8
sqlalchemy>=2.0.0
alembic>=1.12.0
python-dotenv>=1.0.0
asyncpg>=0.28.0  # For PostgreSQL
aiosqlite>=0.19.0  # For SQLite
openai>=1.12.0  # For OpenAI API
fastapi>=0.109.0
uvicorn>=0.27.0 